backend/requirements.txt
flask[async]==2.3.3
flask-cors==4.0.0
livekit==1.0.12
openai==1.12.0
elevenlabs==0.2.26
python-dotenv==1.0.0
requests==2.31.0
websockets==11.0.3
httpx==0.24.1 